backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expert-match-backend"
version = "0.1.0"
description = "FastAPI backend for expert matching and outreach"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "adapters>=1.2.0",
  "cachetools>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.0",
  "python-dateutil>=2.9.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.36",
  "torch==2.6.0+cpu; platform_system == 'Linux'",
  "torch>=2.6.0,<2.7; platform_system != 'Linux'",
  "transformers>=4.49.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "pytest-recording>=0.13.4",
  "ruff>=0.9.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
